Refactor initialization code for a single init_tool() fn in tools
This changes the ToolSpec API to eliminate it and replace it with a
simpler init_fn(url, key) initialization function within each tool
module. The inputs to this are the OpenCTI URL and Key from the cmdline
or environment, and it returns the fn entrypoint for the tool. The
details about how credential management is handled is left up to the
internals of the tool implementation, and the tool entrypoint's function
Annotations are used to seed the LLM's description of the purpose of the
tool.